refactor: abstract supabase with nextjs
