+++ apps/community/src/app/camp/api/sms/route.ts
@@ -4,7 +4,6 @@ import { SolapiMessageService } from "solapi";
 export async function POST(req: Request) {
   // from webhook, get data
   const data: FormbricksWebhookData = await req.json();
-  console.log("🚀 ~ POST ~ data:", data);
   //TODO: 9기 알림 신청에서의 값 즉, 10기 에는 바꿔줘야함
   const to = data?.data?.data?.camyx1dh1a0zzkgq87qo0os8;
   const email = data?.data?.data?.g5m3b6gh9l5rpatm93p7s3pj;
