+++ apps/community/src/features/comment/action/index.ts
@@ -1,11 +1,10 @@
 "use server";
 
-import { revalidatePath } from "next/cache";
-
 import {
   getSupabaseServerClient,
   getUserFromSupabase,
-} from "@/shared/supabase";
+} from "@packages/supabase";
+import { revalidatePath } from "next/cache";
 
 import { grantPointAction } from "../../track-activity/action";
 import { CommentWithAuthor } from "../model/types";
