+++ packages/auth/src/features/auth/model/use-auth.ts
@@ -1,4 +1,4 @@
-import { getSupabaseClient } from "@packages/supabase";
+import { getSupabaseClient } from "@packages/supabase/src/client";
 
 export function useAuth() {
   const supabase = getSupabaseClient({
