--- packages/auth/src/shared/supabase/action.ts
@@ -1,40 +1,0 @@
-import { getSupabaseServerClient } from "./server";
-
-export async function getUserFromSupabase() {
-  const supabase = await getSupabaseServerClient();
-  const {
-    data: { user },
-    error,
-  } = await supabase.auth.getUser();
-
-  if (error) {
-    return null;
-  }
-
-  return user;
-}
-
-export async function getAuthSession() {
-  const supabase = await getSupabaseServerClient();
-  try {
-    const {
-      data: { session },
-      error,
-    } = await supabase.auth.getSession();
-    if (error) throw error;
-    return session;
-  } catch (error) {
-    console.error("Error:", error);
-    return null;
-  }
-}
-
-export async function signOut() {
-  const supabase = await getSupabaseServerClient();
-  try {
-    const { error } = await supabase.auth.signOut();
-    if (error) throw error;
-  } catch (error) {
-    console.error("Error:", error);
-  }
-}
