--- packages/auth/src/shared/supabase/index.ts
@@ -1,2 +1,0 @@
-export * from "./action";
-export * from "./server";
