--- packages/auth/src/shared/supabase/server.ts
@@ -1,7 +1,0 @@
-import { getSupabaseServer } from "@packages/supabase";
-import { cookies } from "next/headers";
-
-export async function getSupabaseServerClient() {
-  const cookieStore = await cookies();
-  return getSupabaseServer(cookieStore);
-}
