+++ packages/supabase/package.json
@@ -12,7 +12,8 @@
     "@packages/env": "workspace:*",
     "@packages/ui": "workspace:*",
     "@supabase/ssr": "^0.4.0",
-    "@supabase/supabase-js": "^2.45.1"
+    "@supabase/supabase-js": "^2.45.1",
+    "next": "^15.3.1"
   },
   "devDependencies": {
     "@repo/eslint-config": "workspace:*",
