+++ packages/supabase/src/index.ts
@@ -2,3 +2,4 @@ export * from "./client";
 export * from "./server";
 export * from "./admin";
 export * from "./types";
+export * from "./action";
