+++ packages/supabase/src/server.ts
@@ -1,7 +1,10 @@
 import { createServerClient } from "@supabase/ssr";
 import { Database } from "./types";
+import { cookies } from "next/headers";
 
-export async function getSupabaseServer(cookieStore: any) {
+export async function getSupabaseServerClient() {
+  const cookieStore = await cookies();
+
   return createServerClient<Database>(
     process.env.NEXT_PUBLIC_SUPABASE_URL!,
     process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
@@ -10,9 +13,9 @@ export async function getSupabaseServer(cookieStore: any) {
         getAll() {
           return cookieStore.getAll();
         },
-        setAll(cookiesToSet: any) {
+        setAll(cookiesToSet) {
           try {
-            cookiesToSet.forEach(({ name, value, options }: any) =>
+            cookiesToSet.forEach(({ name, value, options }) =>
               cookieStore.set(name, value, options)
             );
           } catch {
